requirement2.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
crispy-bootstrap5==0.7
croniter==1.3.7
Deprecated==1.2.13
Django==4.1.2
django-appconf==1.0.5
django-crispy-forms==1.14.0
django-image-cropping==1.7
idna==3.4
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
psycopg2-binary==2.9.4
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
razorpay==1.3.0
redis==4.3.4
requests==2.28.1
rq==1.11.1
rq-scheduler==0.11.0
six==1.16.0
sqlparse==0.4.3
twilio==7.15.0
urllib3==1.26.12
wrapt==1.14.1
